+++ frontends/bmd/src/app_mark_only.test.tsx
@@ -199,7 +199,6 @@ it('MarkOnly flow', async () => {
   await advanceTimersAndPromises();
   fireEvent.click(screen.getByText('Close Polls for Center Springfield'));
   fireEvent.click(screen.getByText('Close VxMark Now'));
-  screen.getByText('Open Polls for Center Springfield');
 
   // Remove card
   card.removeCard();
